chore: move deno/JSR release tooling out of xtask
Revert this branch's xtask changes (deno.json manifest support, JSR links,
the attw/workspaces refactor). The deno.json → JSR publishing logic now lives
in the release-tooling PR (#203) instead, so this branch carries only the
deno packages and their CI. Depends on #203 for xtask to recognize the
deno.json manifests.
